+++ app/renderer/components/progress_spinner.js
@@ -1,0 +1,21 @@
+"use strict";
+
+module.exports = {
+    template: `
+    <div class="progress-spinner">
+        <div class="preloader-wrapper active progress-badge">
+            <div class="spinner-layer spinner-blue-only">
+                <div class="circle-clipper left">
+                    <div class="circle"></div>
+                </div>
+                <div class="gap-patch">
+                    <div class="circle"></div>
+                </div>
+                <div class="circle-clipper right">
+                    <div class="circle"></div>
+                </div>
+            </div>
+        </div>
+    </div>
+    `
+};